Создал абстрактный класс, применил наследование, убрал дубли метдов, перенес классы в папку с компонентами
